+++ ion/gnn/gcn.py
@@ -32,7 +32,6 @@ class GCNConv(Module):
         in_dim: int,
         out_dim: int,
         bias: bool = True,
-        dtype: jnp.dtype = jnp.float32,
         w_init: Initializer = jax.nn.initializers.he_normal(),
         b_init: Initializer = jax.nn.initializers.zeros,
         *,
@@ -40,8 +39,8 @@ class GCNConv(Module):
     ) -> None:
 
         key_w, key_b = jax.random.split(key)
-        self.w = Param(w_init(shape=(in_dim, out_dim), dtype=dtype, key=key_w))
-        self.b = Param(b_init(shape=(out_dim,), dtype=dtype, key=key_b)) if bias else None
+        self.w = Param(w_init(shape=(in_dim, out_dim), key=key_w))
+        self.b = Param(b_init(shape=(out_dim,), key=key_b)) if bias else None
 
     def __call__(
         self,